kayadhu/requirements.txt
amqp==2.3.2
aniso8601==3.0.2
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
bleach==2.1.4
boto3==1.7.79
botocore==1.10.79
braintree==3.47.0
cairocffi==0.9.0
CairoSVG==2.1.3
celery==4.2.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3.1
cssselect2==0.2.1
dataclasses==0.6
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-email-url==0.1.0
Django==2.0.8
django-appconf==1.0.2
django-babel==0.6.2
django-bootstrap4==0.0.6
django-cache-url==3.0.0
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-countries==5.3.1
django-debug-toolbar==1.9.1
django-debug-toolbar-request-history==0.0.7
django-elasticsearch-dsl==0.5.0
django-environ==0.4.5
django-filter==1.1.0
django-fsm==2.6.0
django-graphql-jwt==0.1.10
django-impersonate==1.3
django-js-asset==1.1.0
django-measurement==3.0.0
django-mptt==0.9.1
django-payments==0.12.3
django-phonenumber-field==2.0.0
django-prices==1.0.3
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.1
django-recaptcha==1.4.0
django-redis==4.9.0
django-render-block==0.5
django-rest-auth==0.9.3
django-storages==1.6.6
django-templated-email==2.2.0
django-versatileimagefield==1.9
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djongo==1.2.30
docutils==0.14
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
Faker==0.9.0
freezegun==0.3.10
google-i18n-address==2.3.2
google-measurement-protocol==1.0.0
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.7
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.13.2
jmespath==0.9.3
jsonfield==2.0.2
kombu==4.2.1
Markdown==2.6.11
maxminddb==1.4.1
maxminddb-geolite2==2018.703
measurement==2.0.1
more-itertools==4.3.0
mpmath==1.0.0
multidict==4.3.1
oauthlib==2.1.0
parso==0.3.1
pdfrw==0.4
phonenumberslite==8.9.10
pickleshare==0.7.5
Pillow==5.0.0
pluggy==0.7.1
prices==1.0.0
promise==2.1
prompt-toolkit==2.0.7
psycopg2-binary==2.7.5
purl==1.4
py==1.5.4
pycparser==2.18
Pygments==2.3.1
PyJWT==1.6.4
pymongo==3.7.2
Pyphen==0.9.4
pytest==3.7.1
pytest-django==3.4.1
pytest-mock==1.10.0
pytest-vcr==0.3.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
raven==6.9.0
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
Rx==1.6.1
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
stripe==2.5.0
suds-jurko==0.6
sympy==1.2
text-unidecode==1.2
tinycss2==0.6.1
traitlets==4.3.2
typing==3.6.4
urllib3==1.23
vcrpy==1.13.0
vine==1.1.4
virtualenv==16.0.0
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
wrapt==1.10.11
xmltodict==0.11.0
yarl==1.2.6
